Hotfix
修复了一个弱智但严重的错误（(这个人怎么commit之前都不debug的.jpg）
顺便补充一下上次更新日志忘写的东西：
1.现在save文件只会写入有数据录入的稿件，且改成了续写逻辑
2.为xlsx的header增加了原创情况和引擎两列，以便计录当期相关数据

+++ cvse_新.py
@@ -606,9 +606,10 @@ if with_match:
 if Pres_data.flag != 1:
     load_record(pres_list, f'{_default_dir}/{rank_trans[_rank]}_{_index}_save.csv')
     load_record(pres_list, f'{_default_dir}/{rank_trans[_rank]}_{_index}_save_backup.csv')
-with open(f'{_default_dir}/{rank_trans[_rank]}_{_index}_save.csv', 'w', newline='', encoding='utf-8-sig') as f:
-    f = csv.DictWriter(f, fieldnames=header)
-    f.writeheader()
+if not os.path.exists(f'{_default_dir}/{rank_trans[_rank]}_{_index}_save.csv'):
+    with open(f'{_default_dir}/{rank_trans[_rank]}_{_index}_save.csv', 'w', newline='', encoding='utf-8-sig') as f:
+        f = csv.DictWriter(f, fieldnames=header)
+        f.writeheader()
 place = 1
 new_place = 1
 pres_list.sort(reverse=True)
